Rename RouteCostCalculator
