+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/cost/DefaultTripPattern.java
@@ -5,7 +5,7 @@ import org.opentripplanner.transit.raptor.api.transit.RaptorTripPattern;
 
 public interface DefaultTripPattern extends RaptorTripPattern {
   /**
-   * This is not used by the default calculator, but by the {@link RouteCostCalculator} to
+   * This is not used by the default calculator, but by the {@link PatternCostCalculator} to
    * give unpreferred routes or agencies a generalized-cost penalty.
    */
   Route route();
