+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/cost/PatternCostCalculator.java
@@ -7,7 +7,7 @@ import org.opentripplanner.transit.raptor.api.transit.CostCalculator;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransferConstraint;
 
-public class RouteCostCalculator<T extends DefaultTripSchedule> implements CostCalculator<T> {
+public class PatternCostCalculator<T extends DefaultTripSchedule> implements CostCalculator<T> {
 
   public static final double DEFAULT_ROUTE_RELUCTANCE = 1.0;
   public static final double UNPREFERRED_ROUTE_RELUCTANCE = 2.0;
@@ -16,7 +16,7 @@ public class RouteCostCalculator<T extends DefaultTripSchedule> implements CostC
   private final BitSet unpreferredPatterns;
   private final DoubleFunction<Double> unpreferredCost;
 
-  public RouteCostCalculator(
+  public PatternCostCalculator(
     @Nonnull CostCalculator<T> delegate,
     @Nonnull BitSet unpreferredPatterns,
     @Nonnull DoubleFunction<Double> unpreferredCost
